Edited the code for the description handle event
Changed a necesary option, now the code does not display a text embed in itself, now we search in the mod folder for a .txt file that matches the name of the object. Have in mind that the code has the path for the files located in the appdata for testing purposes, once we launch it to the workshop the code will change the pathing.